[test] UT 전면 수정
