+++ src/main/java/shell/TestScenario.java
@@ -1,0 +1,4 @@
+package shell;
+
+public class TestScenario {
+}
