--- src/test/java/shell/TestScenario.java
@@ -1,72 +1,0 @@
-package shell;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
-
-@ExtendWith(MockitoExtension.class)
-public class TestScenario {
-
-    @Mock
-    TestShell testShell;
-
-
-    private void readCompare(int i, String s) {
-        String result = testShell.read(i);
-        if(!result.equals(s)) fail();
-    }
-
-    @Test
-    void FullWriteAndReadCompare_첫번째_시나리오테스트(){
-        for(int i=0;i<100;i++){
-            when(testShell.read(i)).thenReturn("0xFFFFFFFF");
-        }
-        for(int i=0;i<20;i++){
-            for(int j=0;j<5;j++){
-                testShell.write(i*5+j,"0xFFFFFFFF");
-                readCompare(i*5+j,"0xFFFFFFFF");
-            }
-        }
-        assertTrue(true);
-    }
-
-    @Test
-    void PartialLBAWrite_두번째_시나리오테스트(){
-        for(int i=0;i<150;i++){
-            when(testShell.read(i%5)).thenReturn("0xFFFFFFFF");
-        }
-
-        for(int i=0;i<30;i++){
-            testShell.write(4,"0xFFFFFFFF");
-            testShell.write(0,"0xFFFFFFFF");
-            testShell.write(3,"0xFFFFFFFF");
-            testShell.write(1,"0xFFFFFFFF");
-            testShell.write(2,"0xFFFFFFFF");
-            for(int j=0;j<5;j++){
-                readCompare(j,"0xFFFFFFFF");
-            }
-        }
-
-    }
-
-    @Test
-    void WriteReadAging_세번째_시나리오테스트(){
-        for(int i=0;i<200;i++){
-            when(testShell.read(0)).thenReturn("0xFFFFFFFF");
-            when(testShell.read(99)).thenReturn("0xFFFFFFFF");
-        }
-
-        for(int i=0;i<200;i++){
-            testShell.write(0,"0xFFFFFFFF");
-            readCompare(0,"0xFFFFFFFF");
-            testShell.write(99,"0xFFFFFFFF");
-            readCompare(0,"0xFFFFFFFF");
-        }
-
-    }
-}
